+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSWitchHutsConfig.java
@@ -32,7 +32,8 @@ public class RSWitchHutsConfig
 						.define("blacklistedWitchHutBiomes", " "));
 
 				witchHutsOakMaxChunkDistance = subscriber.subscribe(builder
-						.comment("\n Add Oak Witch Huts to modded Forest biomes that are not birch or dark oak.")
+						.comment("\n Add Oak Witch Huts to modded Forest biomes that are not birch or dark oak.",
+								" 1 for spawning in most chunks and 1001 for none.")
 						.translation("repurposedstructures.config.witch_huts.witchhutsoakmaxchunkdistance")
 						.defineInRange("witchHutsOakMaxChunkDistance", 48, 1, 1001));
 
@@ -42,7 +43,8 @@ public class RSWitchHutsConfig
 						.define("addWitchHutsOakToModdedBiomes", true));
 
 				witchHutsTaigaMaxChunkDistance = subscriber.subscribe(builder
-						.comment("\n Add Taiga Witch Huts to modded Taiga biomes.")
+						.comment("\n Add Taiga Witch Huts to modded Taiga biomes.",
+								" 1 for spawning in most chunks and 1001 for none.")
 						.translation("repurposedstructures.config.witch_huts.witchhutstaigamaxchunkdistance")
 						.defineInRange("witchHutsTaigaMaxChunkDistance", 48, 1, 1001));
 
@@ -52,7 +54,8 @@ public class RSWitchHutsConfig
 						.define("addWitchHutsTaigaToModdedBiomes", true));
 
 				witchHutsGiantTreeTaigaMaxChunkDistance = subscriber.subscribe(builder
-						.comment("\n Add Giant Tree Taiga Witch Huts to modded Giant Tree Taiga biomes.")
+						.comment("\n Add Giant Tree Taiga Witch Huts to modded Giant Tree Taiga biomes.",
+								" 1 for spawning in most chunks and 1001 for none.")
 						.translation("repurposedstructures.config.witch_huts.witchhutsgianttreetaigamaxchunkdistance")
 						.defineInRange("witchHutsGiantTreeTaigaMaxChunkDistance", 48, 1, 1001));
 
@@ -62,7 +65,8 @@ public class RSWitchHutsConfig
 						.define("addWitchHutsGiantTreeTaigaToModdedBiomes", true));
 
 				witchHutsBirchMaxChunkDistance = subscriber.subscribe(builder
-						.comment("\n Add Birch Witch Huts to modded Birch biomes.")
+						.comment("\n Add Birch Witch Huts to modded Birch biomes.",
+								" 1 for spawning in most chunks and 1001 for none.")
 						.translation("repurposedstructures.config.witch_huts.witchhutsbirchmaxchunkdistance")
 						.defineInRange("witchHutsBirchMaxChunkDistance", 48, 1, 1001));
 
@@ -72,7 +76,8 @@ public class RSWitchHutsConfig
 						.define("addWitchHutsBirchToModdedBiomes", true));
 
 				witchHutsDarkForestMaxChunkDistance = subscriber.subscribe(builder
-						.comment("\n Add Dark Forest Witch Huts to modded Dark Forest biomes.")
+						.comment("\n Add Dark Forest Witch Huts to modded Dark Forest biomes.",
+								" 1 for spawning in most chunks and 1001 for none.")
 						.translation("repurposedstructures.config.witch_huts.witchhutsdarkforestmaxchunkdistance")
 						.defineInRange("witchHutsDarkForestMaxChunkDistance", 48, 1, 1001));